fix: cast series to Float64 before drop_nans in adf_test

+++ limen/utils/adf_test.py
@@ -41,7 +41,7 @@ def adf_test(series: pl.Series,
     if not 0 < significance_level < 1:
         raise ValueError(f"significance_level must be between 0 and 1, got {significance_level}")
 
-    values = series.drop_nulls().drop_nans().to_numpy()
+    values = series.cast(pl.Float64).drop_nulls().drop_nans().to_numpy()
     if len(values) == 0:
         return _INCONCLUSIVE
